+++ netlify/functions/club/reviews.ts
@@ -3,7 +3,7 @@ import {
   DetailedReviewResponse,
   Club,
   DateObject,
-} from "@/models";
+} from "@/common/types/models";
 import {
   HandlerEvent,
   HandlerContext,
