+++ package.json
@@ -42,9 +42,9 @@
     "tailwindcss": "^3.0.24",
     "typescript": "~4.5.5",
     "vite": "^2.9.13",
+    "vue-class-component": "8.0.0-rc.1",
     "vue-eslint-parser": "^9.0.1",
-    "vue-tsc": "^0.34.11",
-    "vue-class-component": "8.0.0-rc.1"
+    "vue-tsc": "^0.34.11"
   },
   "eslintConfig": {
     "root": true,
