+++ src/common/components/MoviePosterCard.vue
