+++ src/common/components/NavBar.vue
