+++ src/common/components/VAvatar.vue
