+++ src/common/components/VTable.vue
@@ -75,7 +75,7 @@
 
 <script setup lang="ts">
 import { ref, computed, watch } from 'vue';
-import { Header } from '@/models';
+import { Header } from '@/common/types/models';
 
 interface Props {
   headers: Header[];
