+++ src/data/useRequest.ts
@@ -1,4 +1,4 @@
-import { DataService, FetchConfig } from "@/models";
+import { DataService, FetchConfig } from "@/common/types/models";
 import axios, { AxiosError, AxiosResponse } from "axios";
 import { computed, reactive, ref } from "vue";
 import { useStore } from "vuex";
