+++ src/data/useUser.ts
@@ -1,4 +1,4 @@
-import { CacheDataService, Member } from "@/models";
+import { CacheDataService, Member } from "@/common/types/models";
 import { computed, watch } from "vue";
 import { useStore } from "vuex";
 import { useRequestCache } from "./useRequest";
