+++ src/features/clubs/ClubsView.vue
@@ -26,7 +26,7 @@ import { useStore } from "vuex";
 import clubSvg from "@/assets/images/menu-images/club.svg";
 import { useUser } from "@/data/useUser";
 import { useClub } from "@/data/useClub";
-import { ClubsViewClub, CacheDataService } from "@/models";
+import { ClubsViewClub, CacheDataService } from "@/common/types/models";
 
 const store = useStore();
 const isLoggedIn = computed(() => store.getters["auth/isLoggedIn"]);
