+++ src/features/reviews/components/ReviewCard.vue
@@ -24,8 +24,8 @@
 </template>
 
 <script setup lang="ts">
-  import { Member, ReviewResponse } from '@/models'
-  import MoviePosterCard from "./MoviePosterCard.vue"
+  import { Member, ReviewResponse } from '@/common/types/models'
+  import MoviePosterCard from "../../../common/components/MoviePosterCard.vue"
 
   const { review, members, movieTitle, moviePosterUrl } = defineProps<{
     review: ReviewResponse,
