+++ src/features/reviews/views/ReviewView.vue
@@ -65,7 +65,7 @@
 import { ref, computed, nextTick } from "vue";
 import { useRoute } from "vue-router";
 import AddReviewPrompt from "@/components/SearchPrompt/AddReviewPrompt.vue";
-import { Header } from "@/models";
+import { Header } from "@/common/types/models";
 import { DateTime } from "luxon";
 import { useReview, useSubmitScore } from "@/data/useReview";
 import { useMembers } from "@/data/useClub";
