+++ src/features/reviews/views/ReviewsGalleryView.vue
@@ -39,8 +39,8 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue';
 import { useRoute } from 'vue-router';
-import ReviewCard from '@/components/ReviewCard.vue'
-import LoadingSpinner from '@/components/LoadingSpinner.vue';
+import ReviewCard from '@/features/reviews/components/ReviewCard.vue'
+import LoadingSpinner from '@/common/components/LoadingSpinner.vue';
 import { useDetailedReview } from '@/data/useReview';
 import { useMembers } from '@/data/useClub';
