+++ src/features/statistics/StatisticsView.vue
@@ -61,10 +61,10 @@ import { AgChartsVue } from 'ag-charts-vue3';
 import { useRoute } from "vue-router";
 import axios from "axios";
 import { DateTime } from "luxon";
-import { ReviewResponse, TMDBMovieData, Member, Header } from "@/models";
+import { ReviewResponse, TMDBMovieData, Member, Header } from "@/common/types/models";
 import { useReview } from "@/data/useReview";
 import { useMembers, useClub } from "@/data/useClub";
-import { normalizeArray, loadDefaultChartSettings} from "../util/StatisticsUtils";
+import { normalizeArray, loadDefaultChartSettings} from "../../util/StatisticsUtils";
 
 const route = useRoute();
 const { loading: loadingClub, data: club } = useClub(route.params.clubId as string);
