+++ src/main.ts
@@ -6,13 +6,13 @@ import store from "./store";
 import mdiVue from "mdi-vue/v3";
 import * as mdijs from "@mdi/js";
 
-import VBtn from "@/components/VBtn.vue";
-import VAvatar from "./components/VAvatar.vue";
-import Loading from "@/components/LoadingSpinner.vue"
-import VTable from "@/components/VTable.vue";
+import VBtn from "@/common/components/VBtn.vue";
+import VAvatar from "@/common/components/VAvatar.vue";
+import Loading from "@/common/components/LoadingSpinner.vue"
+import VTable from "@/common/components/VTable.vue";
 import MenuCard from "@/components/MenuCard.vue";
-import VModal from "@/components/VModal.vue";
-import PageHeader from "@/components/PageHeader.vue";
+import VModal from "@/common/components/VModal.vue";
+import PageHeader from "@/common/components/PageHeader.vue";
 import LazyLoad from "./directives/LazyLoad";
 import './assets/styles/tailwind.css'
 import 'animate.css';
